Mudança database

+++ ProjetoBackend/ProjetoBackend/Models/Compra.cs
@@ -15,7 +15,7 @@ namespace ProjetoBackend.Models
         public Fornecedor? Fornecedor { get; set; }
         [Display(Name = "Data da Compra")]
         public DateTime? DataCompra { get; set; } = DateTime.Now;
-        public double? ValorTotal { get; set; } = 0;
+        public decimal? ValorTotal { get; set; } = 0;
 
     }
 }
