+++ ProjetoBackend/ProjetoBackend/Models/Produto.cs
@@ -9,10 +9,10 @@ namespace ProjetoBackend.Models
         public string Nome { get; set; }
         [Required(ErrorMessage = "O Campo Preço do Produto é Obrigatório!")]
         [Display(Name = "Preço")]
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
         [Display(Name = "Estoque")]
         [Required(ErrorMessage = "O Campo Estoque do Produto é Obrigatório!")]
-        public double Estoque { get; set; }
+        public decimal Estoque { get; set; }
 
         /** Relacionamento entre tabelas **/
         /* Chave Estrangeria */
